Extracted /shared folder from /client
