+++ jogging-app/scripts/tools/nextjs/extensions/with-all-extensions.js
@@ -4,13 +4,15 @@ const withFonts = require('./with-fonts');
 const withStyles = require('./with-styles');
 const withFixGraphql = require('./with-fix-graphql');
 const withTranspileModules = require('./with-transpile-local-modules');
+const withResolveAlases = require('./with-resolve-aliases');
 
 module.exports = _.flow([
-  withExcludeModules,
-  withFonts,
-  withStyles,
-  withFixGraphql,
-  withTranspileModules({
-    transpileModules: ['next-routes']
-  })
+    withExcludeModules,
+    withFonts,
+    withStyles,
+    withFixGraphql,
+    withTranspileModules({
+        transpileModules: ['next-routes']
+    }),
+    withResolveAlases
 ]);
