+++ jogging-app/src/shared/guards/auth.guard.js
